Added home page and react router
